tfw you refactor but it still compiles
This commit contains a bunch of cleanup/refactoring I've been putting
off for a while. The biggest piece in here is properly structuring
seperate audio commands, which should make it a ton easier for people
looking to just add or change current sounds.

The second piece was properly doing command handling so you can control
sounds for every type of sound.

also moo for stan